Force passenger related facts on RHEL platform
On RHEL platform we force passenger path until
a package is provided.

+++ configure.sh
@@ -149,9 +149,18 @@ for h in $HOSTS; do
     if [ $h = $(hostname -s) ]; then
         (echo "Configure Puppet environment on ${h} node:"
          mkdir -p /etc/facter/facts.d
-         cat > /etc/facter/facts.d/environment.txt <<EOF
+         if [[ -e /etc/redhat-release ]];
+           then
+           cat > /etc/facter/facts.d/environment.txt <<EOF
+type=${PROF_BY_HOST[$h]}
+apache::mod::passenger::passenger_root="/usr/local/share/gems/gems/passenger-4.0.55"
+apache::mod::passenger::mod_lib_path="/usr/local/share/gems/gems/passenger-4.0.55/buildout/apache2/"
+EOF
+           else
+           cat > /etc/facter/facts.d/environment.txt <<EOF
 type=${PROF_BY_HOST[$h]}
 EOF
+         fi
         n=$(($n + 1)))
     else
         (echo "Configure Puppet environment on ${h} node:"
